+++ plugins/security/api_rolesmapping.go
@@ -24,7 +24,7 @@ func (c rolesmappingClient) Get(ctx context.Context, req *RolesMappingGetReq) (R
 		data RolesMappingGetResp
 		err  error
 	)
-	if data.response, err = c.apiClient.do(ctx, req, &data); err != nil {
+	if data.response, err = c.apiClient.do(ctx, req, &data.RolesMapping); err != nil {
 		return data, err
 	}